Install target platform in local repository
The previous implementation added a temporary project to the reactor,
which caused problems with several other plugins because the project has
never been completely initialized. Instead, we now install the artifact
into the local maven repository. This also avoids temporary files that
are not properly deleted.

The downside of this approach are possible conflicts if two builds use
the same target platform coordinates. This should be avoided.